checking file under 100mb

+++ tasks.py
@@ -5,7 +5,7 @@ from dotenv import load_dotenv
 import os
 import asyncio
 import logging
-from myutils import compress_folder,compress_file,upload_to_filebin,shorten_url,is_folder_less_than_1gb,is_folder_larger_than_2mb,estimate_upload_time,get_ext
+from myutils import compress_folder,compress_file,upload_to_filebin,shorten_url,is_folder_less_than_1gb,is_folder_larger_than_2mb,estimate_upload_time,get_ext,is_file_smaller_than_100mb
 from assistant_file_reader import read_file_with_file_search
 from excelutils import export_to_pdf
 
@@ -139,42 +139,51 @@ def get_summarize(chat_id: str, selected_path: str):
     ext = get_ext(selected_path)
     supported_extension = {"pdf", "doc", "docx", "txt"}
     
-    if ext in supported_extension:
-        logger.info("attempting to send request to openai")
-        prompt=(
-            "Tolong bacakan isi halaman 1 sampai 3 dari file ini."
-            "Jika tidak dapat dibaca, jelaskan penyebabnya secara singkat, tanpa mengajukan pertanyaan."
-        )
-        result = read_file_with_file_search(
-            api_key=OPENAI_APIKEY,
-            file_path=selected_path,
-            prompt=prompt
-        )
-        # dispatcher.utter_message(text=result)
-        logger.info("Got info summarize from openAI %s", result)
-        loop.run_until_complete(
-            bot.send_message(
-                chat_id=chat_id, 
-                text=result,
-                parse_mode="HTML"
+    if is_file_smaller_than_100mb(selected_path):
+        if ext in supported_extension:
+            logger.info("attempting to send request to openai")
+            prompt=(
+                "Tolong bacakan isi halaman 1 sampai 3 dari file ini."
+                "Jika tidak dapat dibaca, jelaskan penyebabnya secara singkat, tanpa mengajukan pertanyaan."
             )
-        )
+            result = read_file_with_file_search(
+                api_key=OPENAI_APIKEY,
+                file_path=selected_path,
+                prompt=prompt
+            )
+            # dispatcher.utter_message(text=result)
+            logger.info("Got info summarize from openAI %s", result)
+            loop.run_until_complete(
+                bot.send_message(
+                    chat_id=chat_id, 
+                    text=result,
+                    parse_mode="HTML"
+                )
+            )
+        else:
+            logger.info(f"attempting to convert {selected_path} to pdf")
+            tmp_file_fullpath = export_to_pdf(selected_path)
+            logger.info("attempting to ask to openai")
+            # perlu optimasi
+            result = read_file_with_file_search(
+                api_key=OPENAI_APIKEY,
+                file_path=tmp_file_fullpath
+            )
+            # dispatcher.utter_message(text=result)
+            logger.info("Got info summarize from openAI %s", result)
+            loop.run_until_complete(
+                bot.send_message(
+                    chat_id=chat_id, 
+                    text=result,
+                    parse_mode="HTML"
+                )
+            )
+            os.remove(tmp_file_fullpath)
     else:
-        logger.info(f"attempting to convert {selected_path} to pdf")
-        tmp_file_fullpath = export_to_pdf(selected_path)
-        logger.info("attempting to ask to openai")
-        # perlu optimasi
-        result = read_file_with_file_search(
-            api_key=OPENAI_APIKEY,
-            file_path=tmp_file_fullpath
-        )
-        # dispatcher.utter_message(text=result)
-        logger.info("Got info summarize from openAI %s", result)
         loop.run_until_complete(
             bot.send_message(
                 chat_id=chat_id, 
-                text=result,
-                parse_mode="HTML"
+                text="Maaf, saya hanya bisa melayani file dibawah `100 MB`",
+                parse_mode="MarkdownV2"
             )
         )
-        os.remove(tmp_file_fullpath)
